Issue #386: move parser.py under deepreg
