+++ deepreg/train.py
@@ -9,8 +9,8 @@ import os
 
 import tensorflow as tf
 
-import deepreg.config.parser as config_parser
 import deepreg.model.optimizer as opt
+import deepreg.parser as config_parser
 from deepreg.model.network.build import build_model
 from deepreg.util import build_dataset, build_log_dir
