+++ test/unit/test_config_parser.py
@@ -8,12 +8,7 @@ import pytest
 import yaml
 from testfixtures import TempDirectory
 
-from deepreg.config.parser import (
-    config_sanity_check,
-    load_configs,
-    save,
-    update_nested_dict,
-)
+from deepreg.parser import config_sanity_check, load_configs, save, update_nested_dict
 
 
 def test_update_nested_dict():
